pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "soccer-vit-linebreak"
version = "0.1.0"
description = "Geometry-labeled soccer line-break pass detection with CNN/ViT explainability"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "numpy>=1.24",
  "pandas>=2.0",
  "pyyaml>=6.0",
  "scikit-learn>=1.3",
]

[project.optional-dependencies]
vision = [
  "torch>=2.2",
  "torchvision>=0.17",
  "timm>=0.9",
  "matplotlib>=3.8",
  "pillow>=10.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
